refactor: crée un composant avec la répartition de la RGCP
