+++ site/source/locales/ui-en.yaml
@@ -596,7 +596,6 @@ design-system:
     open-selector: Open date selector
     prev-month: Previous month
     year: Year
-dont chômage: of which unemployment
 décembre: December
 employeur: employer
 en cas d'accident pro: in the event of a professional accident
