+++ site/source/locales/ui-fr.yaml
@@ -627,7 +627,6 @@ design-system:
     open-selector: Ouvrir le sélecteur de date
     prev-month: Mois précédent
     year: Année
-dont chômage: dont chômage
 décembre: décembre
 employeur: employeur
 en cas d'accident pro: en cas d'accident pro
